Move types protofile, package generation, and custom types to /api/proto/types. Add types.go file to services to simplify this transition.
