--- api/constants.go
@@ -1,34 +1,0 @@
-package api
-
-import (
-	"time"
-)
-
-// TODO: copied over from /lib/defaults to limit imports
-// maybe move to /teleport/constants.go?
-const (
-	// DefaultDialTimeout is a default TCP dial timeout we set for our
-	// connection attempts
-	DefaultDialTimeout = 30 * time.Second
-
-	// HTTPMaxIdleConns is the max idle connections across all hosts.
-	HTTPMaxIdleConns = 2000
-
-	// HTTPMaxIdleConnsPerHost is the max idle connections per-host.
-	HTTPMaxIdleConnsPerHost = 1000
-
-	// HTTPMaxConnsPerHost is the maximum number of connections per-host.
-	HTTPMaxConnsPerHost = 250
-
-	// HTTPIdleTimeout is a default timeout for idle HTTP connections
-	HTTPIdleTimeout = 30 * time.Second
-
-	// KeepAliveCountMax is the number of keep-alive messages that can be sent
-	// without receiving a response from the client before the client is
-	// disconnected. The max count mirrors ClientAliveCountMax of sshd.
-	KeepAliveCountMax = 3
-
-	// ServerKeepAliveTTL is a period between server keep alives,
-	// when servers announce only presence withough sending full data
-	ServerKeepAliveTTL = 60 * time.Second
-)
