+++ api/proto/types/provisioning.go
@@ -1,0 +1,20 @@
+package types
+
+import (
+	fmt "fmt"
+	time "time"
+)
+
+// Expiry returns object expiry setting
+func (p *ProvisionTokenV2) Expiry() time.Time {
+	return p.Metadata.Expiry()
+}
+
+// String returns the human readable representation of a provisioning token.
+func (p ProvisionTokenV2) String() string {
+	expires := "never"
+	if !p.Expiry().IsZero() {
+		expires = p.Expiry().String()
+	}
+	return fmt.Sprintf("ProvisionToken(Roles=%v, Expires=%v)", p.Spec.Roles, expires)
+}
