+++ api/proto/types/remotecluster.go
@@ -1,0 +1,8 @@
+package types
+
+import fmt "fmt"
+
+// String represents a human readable version of remote cluster settings.
+func (c *RemoteClusterV3) String() string {
+	return fmt.Sprintf("RemoteCluster(%v, %v)", c.Metadata.Name, c.Status.Connection)
+}
