+++ api/proto/types/resource.go
@@ -1,0 +1,67 @@
+package types
+
+import (
+	time "time"
+
+	"github.com/jonboulle/clockwork"
+)
+
+// GetMetadata returns object metadata
+func (h *ResourceHeader) GetMetadata() Metadata {
+	return h.Metadata
+}
+
+// GetKind returns resource kind
+func (h *ResourceHeader) GetKind() string {
+	return h.Kind
+}
+
+// GetSubKind returns resource subkind
+func (h *ResourceHeader) GetSubKind() string {
+	return h.SubKind
+}
+
+// GetVersion returns resource version
+func (h *ResourceHeader) GetVersion() string {
+	return h.Version
+}
+
+// GetResourceID returns resource ID
+func (h *ResourceHeader) GetResourceID() int64 {
+	return h.Metadata.ID
+}
+
+// SetResourceID sets resource ID
+func (h *ResourceHeader) SetResourceID(id int64) {
+	h.Metadata.ID = id
+}
+
+// GetName returns the name of the resource
+func (h *ResourceHeader) GetName() string {
+	return h.Metadata.Name
+}
+
+// SetName sets the name of the resource
+func (h *ResourceHeader) SetName(v string) {
+	h.Metadata.SetName(v)
+}
+
+// Expiry returns object expiry setting
+func (h *ResourceHeader) Expiry() time.Time {
+	return h.Metadata.Expiry()
+}
+
+// SetExpiry sets object expiry
+func (h *ResourceHeader) SetExpiry(t time.Time) {
+	h.Metadata.SetExpiry(t)
+}
+
+// SetTTL sets Expires header using current clock
+func (h *ResourceHeader) SetTTL(clock clockwork.Clock, ttl time.Duration) {
+	h.Metadata.SetTTL(clock, ttl)
+}
+
+// SetSubKind sets resource subkind
+func (h *ResourceHeader) SetSubKind(s string) {
+	h.SubKind = s
+}
